# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.CreateGuardianRelationRequest:
    properties:
      guardian_id:
        type: string
      relationship_type:
        type: string
      student_id:
        type: string
    required:
    - guardian_id
    - relationship_type
    - student_id
    type: object
...
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
...
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.GuardianRelationResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      guardian_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      relationship_type:
        type: string
      student_id:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.UpdateGuardianRelationRequest:
    properties:
      is_active:
        type: boolean
      relationship_type:
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: Clean administration API for EduGo
  title: EduGo API Admin New
  version: "1.0"
paths:
...
  /guardian-relations:
    post:
      consumes:
      - application/json
      parameters:
      - description: Guardian relation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.CreateGuardianRelationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.GuardianRelationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a guardian relation
      tags:
      - guardian-relations
  /guardian-relations/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Guardian Relation ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a guardian relation
      tags:
      - guardian-relations
    get:
      consumes:
      - application/json
      parameters:
      - description: Guardian Relation ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.GuardianRelationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a guardian relation by ID
      tags:
      - guardian-relations
    put:
      consumes:
      - application/json
      parameters:
      - description: Guardian Relation ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Guardian relation update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.UpdateGuardianRelationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.GuardianRelationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a guardian relation
      tags:
      - guardian-relations
  /guardians/{guardian_id}/relations:
    get:
      consumes:
      - application/json
      parameters:
      - description: Guardian ID (UUID)
        in: path
        name: guardian_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.GuardianRelationResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get relations for a guardian
      tags:
      - guardian-relations
...
  /menu:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get menu filtered by user permissions
      tags:
      - menu
  /menu/full:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get full menu (admin view)
      tags:
      - menu
  /permissions:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all permissions
      tags:
      - permissions
  /permissions/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Permission ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a permission by ID
      tags:
      - permissions
...
  /students/{student_id}/guardians:
    get:
      consumes:
      - application/json
      parameters:
      - description: Student ID (UUID)
        in: path
        name: student_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.GuardianRelationResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get guardians for a student
      tags:
      - guardian-relations
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and the JWT token. Example: "Bearer
      eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"